add support for sql hint write route only
