+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/hint/SQLHintProperties.java
@@ -15,32 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.hint.sqlhint;
+package org.apache.shardingsphere.infra.hint;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.infra.properties.TypedPropertyKey;
+import org.apache.shardingsphere.infra.properties.TypedProperties;
+
+import java.util.Properties;
 
 /**
- * Typed property key of SQL Hint.
+ * SQL Hint properties.
  */
-@Getter
-@RequiredArgsConstructor
-public enum SQLHintPropertiesKey implements TypedPropertyKey {
+public final class SQLHintProperties extends TypedProperties<SQLHintPropertiesKey> {
     
-    /**
-     * Hint data source name.
-     */
-    DATASOURCE_NAME_KEY("dataSourceName", "", String.class),
-    
-    /**
-     * Whether hint route write data source or not.
-     */
-    WRITE_ROUTE_ONLY_KEY("writeRouteOnly", String.valueOf(Boolean.FALSE), boolean.class);
-    
-    private final String key;
-    
-    private final String defaultValue;
-    
-    private final Class<?> type;
+    public SQLHintProperties(final Properties props) {
+        super(SQLHintPropertiesKey.class, props);
+    }
 }
